chore: upgrade daytona SDK to 0.159.0, merge bridge_assets into bridge
- Bump daytona from >=0.158.0 to >=0.159.0 (archive race condition
  fixes, custom User-Agent headers, datetime serialization fix)
- Merge bridge_assets.py into bridge.py (single consumer, tightly coupled)
