+++ pyproject.toml
@@ -20,7 +20,7 @@ classifiers = [
 ]
 dependencies = [
     "dspy==3.1.3",
-    "daytona>=0.158.0,<1",
+    "daytona>=0.159.0,<1",
     "hydra-core>=1.3,<2",
     "markitdown[all]>=0.1.0,<1",
     "omegaconf>=2.3,<3",
